# Qualcomm AI Engine Direct - Optimize MHA for Gemma3.