--- a/litert/vendors/qualcomm/core/wrappers/tensor_wrapper.h
+++ b/litert/vendors/qualcomm/core/wrappers/tensor_wrapper.h
@@ -112,6 +112,8 @@ class TensorWrapper final {
 
   bool operator==(const TensorWrapper& other) const { return this == &other; }
 
+  bool operator!=(const TensorWrapper& other) const { return this != &other; }
+
   void CloneTo(Qnn_Tensor_t& dst) const;
 
   Qnn_Tensor_t& GetQnnTensor() { return qnn_tensor_; }