… litert/vendors/qualcomm/core/transformation/README.md
… litert/vendors/qualcomm/core/transformation/graph_to_graph_test.cc
… litert/vendors/qualcomm/core/transformation/mha_to_sha.cc